+++ src/conf.ts
@@ -1,0 +1,1 @@
+export const VICTORIA_LOGS_DOCS_HOST= "https://docs.victoriametrics.com";